requirements.txt
fastapi
strawberry-graphql[fastapi]
uvicorn
openai
python-dotenv